Upgrade gradle to 6.5, remove deprecated APIs usage Remove GPars as it's not maintained anymore and switch to java 8 parallel streams Migrate JUnit from 4 to 5 Update jasper version to 6.12.2 migrate the build script from groovy to kotlin

+++ src/main/groovy/com/github/gmazelier/tasks/JasperReportsCompile.groovy
@@ -4,34 +4,34 @@ import net.sf.jasperreports.engine.JasperCompileManager
 import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.tasks.*
-import org.gradle.api.tasks.incremental.IncrementalTaskInputs
+import org.gradle.work.ChangeType
+import org.gradle.work.InputChanges
 
 import java.nio.file.Path
 import java.nio.file.Paths
 
-import static groovyx.gpars.GParsPool.withPool
 
 class JasperReportsCompile extends DefaultTask {
 
 	@InputFiles
 	Iterable<File> classpath
 	@InputDirectory
-	def File srcDir
+	File srcDir
 	@OutputDirectory
-	def File outDir
+	File outDir
 	@Input
-	def String srcExt
+	String srcExt
 	@Input
-	def String outExt
+	String outExt
 	@Input
-	def boolean verbose
+	boolean verbose
 	@Input
-	def boolean useRelativeOutDir
+	boolean useRelativeOutDir
 	@Console
 	def log = getLogger()
 
 	@TaskAction
-	void execute(IncrementalTaskInputs inputs) {
+	void execute(InputChanges inputs) {
 
 		def dependencies = classpath.collect { dependency ->
 			dependency?.toURI()?.toURL()
@@ -39,42 +39,41 @@ class JasperReportsCompile extends DefaultTask {
 		if (verbose) log.lifecycle "Additional classpath: ${dependencies}"
 
 		def compilationTasks = []
-		inputs.outOfDate { change ->
-			if (change.file.name.endsWith(srcExt))
-				compilationTasks << [src: change.file, out: outputFile(change.file), deps: dependencies]
+		inputs.getFileChanges(getInputs().files).forEach{ change ->
+			if(change.changeType != ChangeType.REMOVED){
+				if (change.file.name.endsWith(srcExt))
+					compilationTasks << [src: change.file, out: outputFile(change.file), deps: dependencies]
+			}else{
+				if (verbose) log.lifecycle "Removed file ${change.file.name}"
+				def fileToRemove = outputFile(change.file)
+				fileToRemove.delete()
+			}
 		}
-		inputs.removed { change ->
-			if (verbose) log.lifecycle "Removed file ${change.file.name}"
-			def fileToRemove = outputFile(change.file)
-			fileToRemove.delete()
-		}
+
 
 		def start = System.currentTimeMillis()
 
-		def results = []
-		withPool {
-			results = compilationTasks.collectParallel { task ->
-				def src = task['src'] as File
-				def out = task['out'] as File
-				def deps = task['deps']
+		def results = compilationTasks.parallelStream().collect({task ->
+			def src = task['src'] as File
+			def out = task['out'] as File
+			def deps = task['deps']
 
-				if (verbose) log.lifecycle "Compiling file ${src.name}"
+			if (verbose) log.lifecycle "Compiling file ${src.name}"
 
-				try {
-					// Configure class loader with addtional dependencies
-					ClassLoader originalLoader = Thread.currentThread().getContextClassLoader()
-					URLClassLoader loader = new URLClassLoader(deps as URL[], originalLoader)
-					Thread.currentThread().setContextClassLoader loader
-					// Compile report
-					JasperCompileManager.compileReportToFile src.absolutePath, out.absolutePath
-					// Restore class loader
-					Thread.currentThread().setContextClassLoader originalLoader
-				} catch (any) {
-					return [name: src.name, success: false, exception: any]
-				}
-				[name: src.name, success: true]
+			try {
+				// Configure class loader with addtional dependencies
+				ClassLoader originalLoader = Thread.currentThread().getContextClassLoader()
+				URLClassLoader loader = new URLClassLoader(deps as URL[], originalLoader)
+				Thread.currentThread().setContextClassLoader loader
+				// Compile report
+				JasperCompileManager.compileReportToFile src.absolutePath, out.absolutePath
+				// Restore class loader
+				Thread.currentThread().setContextClassLoader originalLoader
+			} catch (any) {
+				return [name: src.name, success: false, exception: any]
 			}
-		}
+			[name: src.name, success: true]
+		})
 
 		def stop = System.currentTimeMillis()
 
@@ -84,7 +83,7 @@ class JasperReportsCompile extends DefaultTask {
 		log.lifecycle "${results.size()} designs compiled in ${stop - start} ms"
 	}
 
-	def File outputFile(File src) {
+	File outputFile(File src) {
 
 		// local variable so we aren't continuously appending to the outDir
 		def useOutDir = outDir
